Refactored automatic mail deletion
Added a configuration that allows you to delete mails without attachment instantly after the given interval.

Introduced a define for the iteration size of how many mails will be deleted in each circle.
It his highly recommended to not increase it too high or the char server might lag, if you have a lot of mails on the server.